+++ Tests/SwifterSwiftUITests/SwifterSwiftUI/View+ExtensionsTests.swift
@@ -10,32 +10,40 @@ final class ViewExtensionsTests: XCTestCase {
         XCTAssertNoThrow(try anyView.inspect().anyView().emptyView())
     }
 
+    func testEmbedInNavigation() {
+        let anyView = EmptyView().embedInNavigation()
+        XCTAssertNoThrow(try anyView.inspect().navigationView())
+    }
+
     // MARK: Building
     func testIfThen() {
-        let withZIndex = EmptyView().if(true, then: { $0.zIndex(13) })
-        withZIndex.inspect { (view) in
-            XCTAssertNoThrow(try view.anyView().emptyView().zIndex())
-        }
+        let zIndex: Double = 13
 
-        let withoutZIndex = EmptyView().if(false, then: { $0.zIndex(13) })
-        withoutZIndex.inspect { (view) in
-            XCTAssertThrowsError(try view.anyView().emptyView().zIndex())
-        }
+        let withZIndex = EmptyView().if(true, then: { $0.zIndex(zIndex) })
+        XCTAssertNil(withZIndex.value.0)
+        XCTAssertNotNil(withZIndex.value.1)
+        XCTAssertEqual(try withZIndex.value.1!.inspect().emptyView().zIndex(), zIndex)
+
+        let withoutZIndex = EmptyView().if(false, then: { $0.zIndex(zIndex) })
+        XCTAssertNotNil(withoutZIndex.value.0)
+        XCTAssertThrowsError(try withoutZIndex.value.0.inspect().emptyView().zIndex())
+        XCTAssertNil(withoutZIndex.value.1)
     }
 
     func testIfThenElse() {
         let firstIndex = 9.0, secondIndex = 13.0
         let firstView = EmptyView().if(true, then: { $0.zIndex(firstIndex) }, else: { $0.zIndex(secondIndex)})
-        firstView.inspect { (view) in
-            XCTAssertEqual(try view.anyView().emptyView().zIndex(), firstIndex)
-        }
+        XCTAssertNotNil(firstView.value.0)
+        XCTAssertEqual(try firstView.value.0.inspect().emptyView().zIndex(), firstIndex)
+        XCTAssertNil(firstView.value.1)
 
         let secondView = EmptyView().if(false, then: { $0.zIndex(firstIndex) }, else: { $0.zIndex(secondIndex)})
-        secondView.inspect { (view) in
-            XCTAssertEqual(try view.anyView().emptyView().zIndex(), secondIndex)
-        }
+        XCTAssertNil(secondView.value.0)
+        XCTAssertNotNil(secondView.value.1)
+        XCTAssertEqual(try secondView.value.1.inspect().emptyView().zIndex(), secondIndex)
     }
 
+    // MARK: Modifiers
     func testConditionalModifier() {
         let testView = Text("Hello")
         var modifier = InspectableTestModifier()
@@ -47,6 +55,8 @@ final class ViewExtensionsTests: XCTestCase {
             firstExp.fulfill()
         }
         let view = testView.conditionalModifier(true, modifier)
+        XCTAssertNil(view.value.0)
+        XCTAssertNotNil(view.value.1)
         ViewHosting.host(view: view)
         wait(for: [firstExp], timeout: 0.1)
 
@@ -58,6 +68,8 @@ final class ViewExtensionsTests: XCTestCase {
             secondExp.fulfill()
         }
         let secondView = testView.conditionalModifier(false, modifier)
+        XCTAssertNotNil(secondView.value.0)
+        XCTAssertNil(secondView.value.1)
         ViewHosting.host(view: secondView)
         wait(for: [secondExp], timeout: 0.1)
     }
@@ -74,6 +86,8 @@ final class ViewExtensionsTests: XCTestCase {
             firstExp.fulfill()
         }
         let firstView = testView.conditionalModifier(true, thenModifier, elseModifier)
+        XCTAssertNotNil(firstView.value.0)
+        XCTAssertNil(firstView.value.1)
         ViewHosting.host(view: firstView)
         wait(for: [firstExp], timeout: 0.1)
 
@@ -85,10 +99,13 @@ final class ViewExtensionsTests: XCTestCase {
             secondExp.fulfill()
         }
         let secondView = testView.conditionalModifier(false, thenModifier, elseModifier)
+        XCTAssertNil(secondView.value.0)
+        XCTAssertNotNil(secondView.value.1)
         ViewHosting.host(view: secondView)
         wait(for: [secondExp], timeout: 0.1)
     }
 
+    // MARK: Animations
     func testAnimateOnAppear() {
         let testView = Text("Hello")
 
@@ -130,12 +147,6 @@ final class ViewExtensionsTests: XCTestCase {
         ("testAnimateOnAppear", testAnimateOnAppear),
         ("testAnimateOnDisappear", testAnimateOnDisappear)
     ]
-}
-
-private struct TestModifier: ViewModifier {
-    func body(content: Self.Content) -> some View {
-        content.onAppear()
-    }
 }
 
 private struct InspectableTestModifier: ViewModifier {