+++ CHANGELOG.md
@@ -18,6 +18,28 @@ The changelog for **SwifterSwiftUI**. Also see the [releases](https://github.com
 
 ---
 
-## [v0.0.0](https://github.com/asam139/SwifterSwiftUI/releases/tag/0.0.0)
+## [v0.3.0](https://github.com/asam139/SwifterSwiftUI/releases/tag/0.3.0)
+
+### Added
+- **Image**:
+    - `init(_:, defaultImage:)`,  `init(_:, defaultSystemImage:)`.  [#2](https://github.com/asam139/Steps/pull/2) by [asam139](https://github.com/asam139)
+- **Color**: 
+    - `init?(hex:)`. [#2](https://github.com/asam139/Steps/pull/2) by [asam139](https://github.com/asam139)
+- **View**:
+    - `embedInNavigation()`.  [#2](https://github.com/asam139/Steps/pull/2) by [asam139](https://github.com/asam139) 
+
+---
+
+## [v0.2.1](https://github.com/asam139/SwifterSwiftUI/releases/tag/0.2.1)
+
+### Added
+- **LinearGradient**:
+    - `eraseToAnyView()`, `if(_:, then:)`, `if(_:, then:, else:)`, `conditionalModifier(_:, modifier:)`, `conditionalModifier(_:, trueModifier:, falseModifier:)`, `animateOnAppear(using:, action:)`, `animateOnDisappear(using:, action:)`.  [#1](https://github.com/asam139/Steps/pull/1) by [asam139](https://github.com/asam139)
+
+- **LinearGradient**:
+    - `init(_:, startPoint:, endPoint)`.  [#1](https://github.com/asam139/Steps/pull/1) by [asam139](https://github.com/asam139)
+
+- **Utils**:
+    - `ifLet(value:, then:)`, `ifLet(value:, then:, else:)`.  [#1](https://github.com/asam139/Steps/pull/1) by [asam139](https://github.com/asam139)
 
 ---